모노레포 실제 서버와 클라이언트 - 34
1. 클라이언트
 - board, calendar에 간단한 데이터 검증 기능 추가
 - ddd 구조는 도메인별 컴포넌트가 위치해야 하는데 해당 구조는 그렇지 않음!
 - 도메인별 컴포넌트 위치를 바꿀 필요가 있고, 공통으로 쓰는(input, 모달 같은) 컴포넌트만 글로벌 components 디렉토리에 위치하는 게 맞음!
 - 이건 나중에...
 - 가끔 프로젝트를 구동하고 나면 오류가 나는 경우가 있음(로그인이 안됨, 확인 필요)

+++ apps/client/app/components/BoardCRUDComponent.tsx
@@ -28,6 +28,13 @@ const BoardCRUDComponent = () => {
   // ✅ 삽입 (useMutation 사용)
   const insertMutation = useMutation({
     mutationFn: async () => {
+      if (!title.trim() || !content.trim()) {
+        alert("제목 또는 내용을 입력하세요.");
+        throw new Error("제목 또는 내용을 입력하세요."); //에러를 발생해야 입력 전 상태를 기억함! 단순 return인 경우 초기화
+      }
+      if (!confirm("등록하시겠습니까?")) {
+        throw new Error("stop");
+      }
       await useBoardService.insertData({
         title,
         content,
@@ -44,6 +51,13 @@ const BoardCRUDComponent = () => {
   // ✅ 수정 (useMutation 사용)
   const updateMutation = useMutation({
     mutationFn: async (item: Board) => {
+      if (!title.trim() || !content.trim()) {
+        alert("제목 또는 내용을 입력하세요.");
+        return; //에러를 발생해야 입력 전 상태를 기억함! 단순 return인 경우 초기화
+      }
+      if (!confirm("수정하시겠습니까?")) {
+        throw new Error("stop");
+      }
       await useBoardService.updateData(item);
     },
     onMutate: () => {
@@ -67,6 +81,9 @@ const BoardCRUDComponent = () => {
   // ✅ 삭제 (useMutation 사용)
   const deleteMutation = useMutation({
     mutationFn: async (id: string) => {
+      if (!confirm("삭제하시겠습니까?")) {
+        throw new Error("stop");
+      }
       await useBoardService.deleteData(id);
     },
     onSuccess: () => {
